+++ test/fixtures/async-chained-bundle.js
@@ -1,0 +1,5 @@
+globalThis.render = function(data) {
+  return Promise.resolve().then(function() {
+    return '<h1>async-chained</h1>';
+  });
+};