+++ test/fixtures/async-reject-bundle.js
@@ -1,0 +1,3 @@
+globalThis.render = function(data) {
+  return Promise.reject(new Error('async-rejection'));
+};